split data handling and syncing to different file in p2p
