+++ go.mod
@@ -1,7 +1,7 @@
 module github.com/sprawl/sprawl
 
 require (
-	cloud.google.com/go v0.44.1 // indirect
+	cloud.google.com/go v0.44.1
 	github.com/alecthomas/template v0.0.0-20190718012654-fb15b899a751 // indirect
 	github.com/alecthomas/units v0.0.0-20190717042225-c3de453c63f4 // indirect
 	github.com/bouk/monkey v1.0.1
